requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
arrow==1.1.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.3
cssselect==1.1.0
cycler==0.11.0
docopt==0.6.2
fake-useragent==0.1.11
fastapi==0.63.0
fonttools==4.38.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==3.2
importlib-metadata==4.12.0
Jinja2==3.0.1
jinja2-time==0.2.0
kiwisolver==1.4.4
loguru==0.5.3
lxml==4.9.1
MarkupSafe==2.0.1
matplotlib==3.5.0
multidict==5.1.0
nb-cli==0.4.2
nonebot-adapter-cqhttp==2.0.0a13
nonebot2==2.0.0a13.post1
numpy==1.24.2
packaging==23.0
parse==1.19.0
Pillow==8.3.1
pipreqs==0.4.11
poyo==0.5.0
prompt-toolkit==1.0.14
pydantic==1.8.2
pyee==8.2.2
pyfiglet==0.8.post1
Pygments==2.9.0
pygtrie==2.4.2
PyInquirer==1.0.3
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
requests-html==0.10.0
rfc3986==1.5.0
setuptools-scm==7.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.13.6
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.7.2
tqdm==4.64.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.13.4
w3lib==1.22.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.3
win32-setctime==1.0.3
yarg==0.1.9
yarl==1.6.3
zipp==3.8.1
